Parse canonical PHP array keys as 64-bit, and retract a false safety claim
isCanonicalInteger() admits PHP's full signed 64-bit key range, but encode()
parsed the key back with Integer.parseInt(). A key it had just accepted --
2147483648, a legal department or branch name -- threw NumberFormatException,
so dashboard/stats.php answered 500 where PHP returns the object-valued map.

The workforce-planning javadoc separately claimed the unscoped department join
was "safe today because workforce_planning rows are themselves company-scoped".
That is wrong: department_id carries no foreign key, so a company's own row can
point at another tenant's department and disclose its name and headcount. The
query still matches stats.php exactly (D-058); only the claim about it changes.

+++ backend/src/main/java/com/workin/legacy/wire/LegacyPhpArrayJson.java
@@ -60,9 +60,13 @@ public final class LegacyPhpArrayJson {
 			// close, so honouring the cast is the faithful choice here.
 			return map;
 		}
-		int expected = 0;
+		// `long`, not `int`: isCanonicalInteger() admits PHP's full signed
+		// 64-bit key range, so parsing back with Integer.parseInt() would throw
+		// on a key it had just accepted -- a department named `2147483648` is a
+		// legal map key here and would 500 instead of encoding.
+		long expected = 0;
 		for (String key : map.keySet()) {
-			if (!isCanonicalInteger(key) || Integer.parseInt(key) != expected) {
+			if (!isCanonicalInteger(key) || Long.parseLong(key) != expected) {
 				return map;
 			}
 			expected++;
